Add remove book from Api
