+++ src/redux/configureStore.js
@@ -1,7 +1,7 @@
 import { createStore, combineReducers, applyMiddleware } from 'redux';
 import thunk from 'redux-thunk';
 import logger from 'redux-logger';
-import booksReducer from './books/bookApi';
+import booksReducer from './books/book';
 
 const reducer = combineReducers({
   booksReducer,
